requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-braces==1.15.0
django-cors-headers==3.10.1
django-oauth-toolkit==1.6.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
